move: Input 컴포넌트 design 폴더 -> common 폴더
