+++ src/components/common/Input/Input.tsx
@@ -1,8 +1,8 @@
 import type { Size } from '@/types/temp';
 import type { ComponentPropsWithRef, ForwardedRef, ReactElement } from 'react';
 import React, { forwardRef } from 'react';
-import ErrorMessage from '../ErrorMessage/ErrorMessage';
-import Label from '../Label/Label';
+import ErrorMessage from '../../design/ErrorMessage/ErrorMessage';
+import Label from '../../design/Label/Label';
 import {
   getInputStyling,
   getSizeStyling,
