+++ src/components/common/InputComment/InputNewComment/InputNewComment.tsx
@@ -3,7 +3,7 @@ import React, { ChangeEvent, KeyboardEvent } from 'react';
 import { css } from '@emotion/react';
 import { useCreateComment } from '@/hooks/comment/useCreateComment';
 import Button from '@/components/common/Button/Button';
-import Input from '@/components/design/Input/Input';
+import Input from '@/components/common/Input/Input';
 
 interface InputNewCommentProps {
   value: string;
