+++ src/components/common/InputComment/InputNewReply/InputNewReply.tsx
@@ -3,7 +3,7 @@ import React, { ChangeEvent, KeyboardEvent } from 'react';
 import { css } from '@emotion/react';
 
 import Button from '@/components/common/Button/Button';
-import Input from '@/components/design/Input/Input';
+import Input from '@/components/common/Input/Input';
 import { useCreateReply } from '@/hooks/comment/useCreateReply';
 
 interface InputNewReplyProps {
