+++ src/components/common/InputsUserInfo/InputNickname.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { ChangeEvent, useEffect } from 'react';
 import { css } from '@emotion/react';
-import Input from '../../design/Input/Input';
+import Input from '../Input/Input';
 import Button from '../Button/Button';
 import { useCheckNickname } from '../../../hooks/common/inputsUserInfo/useCheckNickname';
