+++ src/components/common/InputsUserInfo/InputPwCheck.tsx
@@ -2,7 +2,7 @@
 import { css } from '@emotion/react';
 import React, { ChangeEvent, useEffect } from 'react';
 import { useCheckPasswordCheck } from '../../../hooks/common/inputsUserInfo/useCheckPasswordCheck';
-import Input from '../../design/Input/Input';
+import Input from '../Input/Input';
 
 interface InputPwCheckProps {
   value: string;
